+++ src/util/interface/SxOperations.ts
@@ -1,0 +1,8 @@
+import { SxType } from "../type/SxType";
+
+/**
+ * Structure for an object that holds {@link https:// Mantine sx} objects.
+ */
+export interface SxOperations {
+    [key: string]: SxType;
+}