+++ src/util/style/index.ts
@@ -1,3 +1,0 @@
-export * from "./merge";
-export * from "./parseStyles";
-export * from "./parseSx";