+++ src/util/type/Class.ts
@@ -1,1 +1,4 @@
+/**
+ * Generic class type.
+ */
 export type Class<T> = new (...args: any[]) => T;