+++ src/util/type/index.ts
@@ -1,4 +1,0 @@
-export * from "./Class";
-export * from "./Optional";
-export * from "./StyleType";
-export * from "./SxType";